Extract approvable logic to a concern
