+++ spec/models/category_spec.rb
@@ -13,47 +13,8 @@ RSpec.describe Category, type: :model do
     it { is_expected.to validate_uniqueness_of(:color).case_insensitive }
   end
 
-  describe '.approved' do
+  include_examples 'approvable' do
     let!(:approved) { create(:category, :approved) }
-
-    before do
-      create(:category, :disapproved)
-    end
-
-    it 'returns one item' do
-      expect(described_class.approved).to have(1).item
-    end
-
-    it 'returns only approved items' do
-      expect(described_class.approved).to match_array(approved)
-    end
-  end
-
-  describe '#approve' do
-    subject { create(:category, :disapproved) }
-    let(:current_user) { create(:user) }
-    let(:now) { Time.current }
-
-    before do
-      Timecop.freeze(now)
-    end
-
-    it 'approves the category' do
-      expect do
-        subject.approve!(current_user)
-      end.to change(subject, :approved).to(true)
-    end
-
-    it 'updates the approved_by to the current_user' do
-      expect do
-        subject.approve!(current_user)
-      end.to change(subject, :approved_by).to(current_user)
-    end
-
-    it 'updates the approved_at' do
-      expect do
-        subject.approve!(current_user)
-      end.to change(subject, :approved_at).to(now)
-    end
+    let!(:disapproved) { create(:category, :disapproved) }
   end
 end
